* commit 'a18e400833c75f0558894965a30bbe37b5143dc0':
  Use long as file size instead of int while using OPP.

+++ src/com/android/bluetooth/opp/BluetoothOppObexServerSession.java
@@ -268,7 +268,9 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
         ContentValues values = new ContentValues();
 
         values.put(BluetoothShare.FILENAME_HINT, name);
-        values.put(BluetoothShare.TOTAL_BYTES, length.intValue());
+
+        values.put(BluetoothShare.TOTAL_BYTES, length);
+
         values.put(BluetoothShare.MIMETYPE, mimeType);
 
         values.put(BluetoothShare.DESTINATION, destination);
